rename and move suite_decoupled_test into testutil.
