+++ controllers/pipelines/pipeline/pipeline_controller_decoupled_test.go
@@ -8,9 +8,9 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/sky-uk/kfp-operator/apis"
 	pipelinesv1 "github.com/sky-uk/kfp-operator/apis/pipelines/v1alpha6"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/testutil"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines"
 	providers "github.com/sky-uk/kfp-operator/argo/providers/base"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/testutil"
 	v1 "k8s.io/api/core/v1"
 )
