+++ controllers/pipelines/run/run_workflow_factory.go
@@ -5,10 +5,10 @@ import (
 
 	config "github.com/sky-uk/kfp-operator/apis/config/v1alpha6"
 	pipelinesv1 "github.com/sky-uk/kfp-operator/apis/pipelines/v1alpha6"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines/runconfiguration"
 	"github.com/sky-uk/kfp-operator/argo/common"
 	providers "github.com/sky-uk/kfp-operator/argo/providers/base"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines/runconfiguration"
 )
 
 type RunDefinitionCreator struct {
