+++ controllers/pipelines/runconfiguration/runconfiguration_controller.go
@@ -18,8 +18,8 @@ import (
 
 	pipelinesv1 "github.com/sky-uk/kfp-operator/apis/pipelines/v1alpha6"
 	common "github.com/sky-uk/kfp-operator/controllers/pipelines"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/statehandler"
 	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/logging"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/statehandler"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
