+++ controllers/pipelines/runconfiguration/runconfiguration_controller_decoupled_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/onsi/gomega/types"
 	"github.com/sky-uk/kfp-operator/apis"
 	pipelinesv1 "github.com/sky-uk/kfp-operator/apis/pipelines/v1alpha6"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/testutil"
 	"github.com/sky-uk/kfp-operator/argo/common"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/testutil"
 )
 
 var _ = Describe("RunConfiguration controller k8s integration", Serial, func() {
