+++ controllers/pipelines/runschedule/runschedule_controller.go
@@ -7,8 +7,8 @@ import (
 	config "github.com/sky-uk/kfp-operator/apis/config/v1alpha6"
 	pipelinesv1 "github.com/sky-uk/kfp-operator/apis/pipelines/v1alpha6"
 	"github.com/sky-uk/kfp-operator/controllers/pipelines"
-	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/statehandler"
 	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/logging"
+	"github.com/sky-uk/kfp-operator/controllers/pipelines/internal/statehandler"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
